requirements.txt
asgiref==3.5.2
Django==4.1.3
fontawesomefree==6.2.0
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.3

djangorestframework~=3.14.0